telegram-bot/requirements.txt
aiogram==3.22.0
supabase==2.0.0
python-dotenv==1.0.0
aiohttp==3.9.0
Pillow==10.0.0
python-slugify==8.0.0

